+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 **3.0.3- 08/20/24**
 
  - Use external script to find matching dependency branches
+ - Fix bug when excluding all but one causes of disability from results
 
 **3.0.2- 08/19/24**